+++ Ajedrez/src/Peon.h
@@ -1,9 +1,13 @@
 #pragma once
-#include "stdio.h"
 
-using namespace std;
+#include <vector>
+#include "Pieza.h"
+#include "Tablero.h"
 
-class Peon {
-	
+class ClassPeon : public ClassPieza{  
 
-}; 
+public: 
+	ClassPeon(Color color, Vector2D posicion);
+	vector<Vector2D> obtenerMovimientosPosibles(const ClassTablero& tablero) const;
+
+};
